refactor: reorganize code
