+++ .env.example
@@ -6,8 +6,7 @@
 # secrets in it. If you are cloning this repo, create a copy of this file named
 # ".env" and populate it with your secrets.
 
-# When adding additional environment variables, the schema in "/src/env/server.ts"
-# and "/src/env/client.ts" should be updated accordingly.
+# When adding additional environment variables, the schema in "/src/lib/env.ts"
 
 # Docker
 PGUSER="postgres"
@@ -15,11 +14,8 @@ PGPASSWORD="postgres"
 PGDATABASE="postgres"
 PGHOST="localhost:5432"
 
-# Prisma
-DATABASE_URL="postgresql://${PGUSER}:${PGPASSWORD}@${PGHOST}/${PGDATABASE}?schema=public"
-DIRECT_URL="postgresql://${PGUSER}:${PGPASSWORD}@${PGHOST}/${PGDATABASE}?schema=public"
-
 # Server
+DATABASE_URL="postgresql://${PGUSER}:${PGPASSWORD}@${PGHOST}/${PGDATABASE}?schema=public"
 NODE_ENV="development"
 
 # Client
