+++ prisma/schema.prisma
@@ -5,10 +5,4 @@ generator client {
 datasource db {
   provider  = "postgresql"
   url       = env("DATABASE_URL")
-  directUrl = env("DIRECT_URL")
-}
-
-model Test {
-  id   String @id @default(cuid())
-  name String
 }
