+++ src/app/_components/theme-toggler.tsx
@@ -3,7 +3,7 @@
 import { Moon, Sun } from 'lucide-react';
 import { useTheme } from 'next-themes';
 
-import { Button } from '@/components/ui/button';
+import { Button } from './ui/button';
 
 export function ThemeToggler() {
   const { systemTheme, theme, setTheme } = useTheme();
