+++ src/app/_components/ui/form.tsx
@@ -10,8 +10,8 @@ import {
   useFormContext,
 } from 'react-hook-form';
 
-import { Label } from '@/components/ui/label';
-import { cn } from '@/lib/utils';
+import { Label } from '@/app/_components/ui/label';
+import { cn } from '@/lib/shadcn';
 
 const Form = FormProvider;
