+++ src/app/_components/ui/input.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import { cn } from '@/lib/utils';
+import { cn } from '@/lib/shadcn';
 
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
 export interface InputProps
