+++ src/app/auth/confirmation/page.tsx
@@ -1,5 +1,5 @@
-import { ConfirmationForm } from '@/components/auth/confirmation-form';
-import { FormWrapper } from '@/components/auth/form-wrapper';
+import { ConfirmationForm } from '@/app/_components/auth/confirmation-form';
+import { FormWrapper } from '@/app/_components/auth/form-wrapper';
 
 export default function AuthConfirmationPage() {
   return (
