+++ src/app/auth/layout.tsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 interface AuthLayoutProps {
   children: React.ReactNode;
 }
