+++ src/app/layout.tsx
@@ -2,10 +2,11 @@ import { GeistMono } from 'geist/font/mono';
 import { GeistSans } from 'geist/font/sans';
 import type { Metadata } from 'next';
 
-import { Header } from '@/components/header';
-import { ThemeProvider } from '@/components/theme-provider';
-import { cn } from '@/lib/utils';
-import '@/styles/globals.css';
+import '@/app/_styles/globals.css';
+import { cn } from '@/lib/shadcn';
+
+import { Header } from './_components/header';
+import { ThemeProvider } from './_components/theme-provider';
 
 export const metadata: Metadata = {
   title: 'Startpage',
