+++ src/app/not-found.tsx
@@ -1,7 +1,7 @@
 import { ArrowLeft } from 'lucide-react';
 import Link from 'next/link';
 
-import { Button } from '@/components/ui/button';
+import { Button } from './_components/ui/button';
 
 export default function NotFound() {
   return (
