+++ src/lib/db.ts
@@ -1,6 +1,6 @@
 import { PrismaClient } from '@prisma/client';
 
-import { env } from './env.server';
+import { env } from './env';
 
 const globalForPrisma = globalThis as unknown as {
   prisma: PrismaClient | undefined;
