+++ src/lib/env.ts
@@ -3,11 +3,14 @@ import { z } from 'zod';
 
 export const env = createEnv({
   server: {
+    DATABASE_URL: z.string().url(),
     NODE_ENV: z
       .enum(['development', 'test', 'production'])
       .default('development'),
   },
+  client: {},
   runtimeEnv: {
+    DATABASE_URL: process.env.DATABASE_URL,
     NODE_ENV: process.env.NODE_ENV,
   },
   skipValidation: !!process.env.SKIP_ENV_VALIDATION,
